Fixes the failing test from previous commit.
Tracking this one down wasn't obvious or easy.

Here's how I figured it out. I added a lot of debugging code and
noticed that 'setup' was being called twice. That wasn't good;
there should have only been one Test::Unit::TestCase executing.

As I dug in deeper, I found that the following code block was
getting executed with 'self' set to TestSharedLifecycle:

  test "battery should be ok" do
    assert_equal :ok, @battery
  end

As a result, this test was running independently of the before
block I had setup:

  before do
    @battery = :ok
    @engine = :off
  end

When I took at look at Context::SharedBehavior.included, it was
clear that something was wrong:

  def included(arg)
    @_behavior.call
  end

This is the reason that @_behavior (a stored block) was being
run with 'self' set to TestSharedLifecycle -- since that was
the original object in which the block was defined.

This was easy to fix with a little help from class_eval.

Side note: it is a little unusual that this code is using the
private instance method 'Module#included' -- it is more common to
create your own class and do 'def self.included'.